refactor(search page): move search to homepage
